refactor: 인증 관련 모델 구조 개선
- LoginData 모델 추가: 로그인 요청 데이터 분리
- TokenData 모델 추가: 토큰 데이터 명시적 관리
- UserModel 삭제: 불필요한 모델 제거
- ProfileModel 수정: 프로필 정보 구조 개선
