+++ lib/models/login_data.dart
@@ -1,0 +1,17 @@
+// ignore_for_file: invalid_annotation_target
+
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+part 'login_data.freezed.dart';
+part 'login_data.g.dart';
+
+@freezed
+class LoginData with _$LoginData {
+  factory LoginData({
+    required String accessToken,
+    required String refreshToken,
+  }) = _LoginData;
+
+  factory LoginData.fromJson(Map<String, dynamic> json) =>
+      _$LoginDataFromJson(json);
+}
