+++ lib/models/login_data.freezed.dart
@@ -1,0 +1,170 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'login_data.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+LoginData _$LoginDataFromJson(Map<String, dynamic> json) {
+  return _LoginData.fromJson(json);
+}
+
+/// @nodoc
+mixin _$LoginData {
+  String get accessToken => throw _privateConstructorUsedError;
+  String get refreshToken => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $LoginDataCopyWith<LoginData> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $LoginDataCopyWith<$Res> {
+  factory $LoginDataCopyWith(LoginData value, $Res Function(LoginData) then) =
+      _$LoginDataCopyWithImpl<$Res, LoginData>;
+  @useResult
+  $Res call({String accessToken, String refreshToken});
+}
+
+/// @nodoc
+class _$LoginDataCopyWithImpl<$Res, $Val extends LoginData>
+    implements $LoginDataCopyWith<$Res> {
+  _$LoginDataCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? accessToken = null,
+    Object? refreshToken = null,
+  }) {
+    return _then(_value.copyWith(
+      accessToken: null == accessToken
+          ? _value.accessToken
+          : accessToken // ignore: cast_nullable_to_non_nullable
+              as String,
+      refreshToken: null == refreshToken
+          ? _value.refreshToken
+          : refreshToken // ignore: cast_nullable_to_non_nullable
+              as String,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$LoginDataImplCopyWith<$Res>
+    implements $LoginDataCopyWith<$Res> {
+  factory _$$LoginDataImplCopyWith(
+          _$LoginDataImpl value, $Res Function(_$LoginDataImpl) then) =
+      __$$LoginDataImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({String accessToken, String refreshToken});
+}
+
+/// @nodoc
+class __$$LoginDataImplCopyWithImpl<$Res>
+    extends _$LoginDataCopyWithImpl<$Res, _$LoginDataImpl>
+    implements _$$LoginDataImplCopyWith<$Res> {
+  __$$LoginDataImplCopyWithImpl(
+      _$LoginDataImpl _value, $Res Function(_$LoginDataImpl) _then)
+      : super(_value, _then);
+
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? accessToken = null,
+    Object? refreshToken = null,
+  }) {
+    return _then(_$LoginDataImpl(
+      accessToken: null == accessToken
+          ? _value.accessToken
+          : accessToken // ignore: cast_nullable_to_non_nullable
+              as String,
+      refreshToken: null == refreshToken
+          ? _value.refreshToken
+          : refreshToken // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$LoginDataImpl implements _LoginData {
+  _$LoginDataImpl({required this.accessToken, required this.refreshToken});
+
+  factory _$LoginDataImpl.fromJson(Map<String, dynamic> json) =>
+      _$$LoginDataImplFromJson(json);
+
+  @override
+  final String accessToken;
+  @override
+  final String refreshToken;
+
+  @override
+  String toString() {
+    return 'LoginData(accessToken: $accessToken, refreshToken: $refreshToken)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$LoginDataImpl &&
+            (identical(other.accessToken, accessToken) ||
+                other.accessToken == accessToken) &&
+            (identical(other.refreshToken, refreshToken) ||
+                other.refreshToken == refreshToken));
+  }
+
+  @JsonKey(ignore: true)
+  @override
+  int get hashCode => Object.hash(runtimeType, accessToken, refreshToken);
+
+  @JsonKey(ignore: true)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$LoginDataImplCopyWith<_$LoginDataImpl> get copyWith =>
+      __$$LoginDataImplCopyWithImpl<_$LoginDataImpl>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$LoginDataImplToJson(
+      this,
+    );
+  }
+}
+
+abstract class _LoginData implements LoginData {
+  factory _LoginData(
+      {required final String accessToken,
+      required final String refreshToken}) = _$LoginDataImpl;
+
+  factory _LoginData.fromJson(Map<String, dynamic> json) =
+      _$LoginDataImpl.fromJson;
+
+  @override
+  String get accessToken;
+  @override
+  String get refreshToken;
+  @override
+  @JsonKey(ignore: true)
+  _$$LoginDataImplCopyWith<_$LoginDataImpl> get copyWith =>
+      throw _privateConstructorUsedError;
+}
